requirements.txt
pandas
numpy<2.0
scikit-learn
lightgbm
tqdm
alpaca-py
python-dotenv
blinker>=1.4 
gdown
matplotlib
pytorch-tcn
polars